% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNArray_pipeline.R
\name{RNArray}
\alias{RNArray}
\title{RNArray pipeline for aligning RNA-seq reads to microarray probesets and making distributions comparable}
\usage{
RNArray(rna_seq_files, array_files, ref_file, paired_end = T,
  frag_length = NA, cores = 1, kallisto_cmd = "kallisto",
  tech_prefix = "AFFX", out_dir = NA)
}
\arguments{
\item{rna_seq_files}{The RNA-seq FASTQ-files of interest (they must have one of the extensions fastq.gz, fq.gz, fqz, fq, or fastq). If your files are paired-end, they should follow the naming scheme: PairA_1.fq, PairA_2.fq, PairB_1.fq, PairB_2.fq etc.}

\item{array_files}{The microarray CEL-files of interest (they must have the extension CEL).}

\item{ref_file}{The name of the fasta file (or kallisto .idx file) containing the reference sequences for you array of interest.}

\item{paired_end}{LOGICAL, either true (RNA-seq reads are paired-end, default) or false (reads are single-end).}

\item{frag_length}{Fragment length of the RNA-seq. Required when paired_end = F, optional when paired_end = T.}

\item{cores}{(Optional) Number of cores to use - multithreading significantly increases the speed (default = 1).}

\item{kallisto_cmd}{(Optional) STRING giving full command to use to call kallisto, if simply typing "kallisto" at the command line does not give the required version of kallisto or does not work. Default is simply "kallisto". If used, this argument should give the full path to the desired kallisto binary.}

\item{tech_prefix}{(Optional) STRING giving the prefix of technical rows on the applied platform - for HGU-133 Plus 2.0 it's 'AFFX'.}

\item{out_dir}{(Optional) STRING giving an output directory. If not specified, the results will be saved in a sub-directory (RNArray_output) in the directory containing the reference file.}
}
\value{
Quantile normalized and batch corrected RNA-seq and microarray data
}
\description{
RNArray pipeline for aligning RNA-seq reads to microarray probesets and making distributions comparable
}
\examples{
RNArray(c('Desktop/FASTQ/FileA.fq.gz', 'Desktop/FASTQ/FileB.fq.gz'), c('Desktop/CEL/FileX.CEL', 'Desktop/CEL/FileY.CEL'), 'Desktop/my_targets.fa', paired_end = T, cores = 4)
RNArray(c('Desktop/FASTQ/FileA.fq.gz', 'Desktop/FASTQ/FileB.fq.gz'), c('Desktop/CEL/FileX.CEL', 'Desktop/CEL/FileY.CEL'), 'Desktop/my_targets.fa', paired_end = F, frag_length = 89, cores = 2)

}
\author{
Christina Bligaard Pedersen and Lars Rønn Olsen
}
