% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNArray_UI.R
\name{RNArray_UI}
\alias{RNArray_UI}
\title{RNArray shiny-based graphical user interface for file selection and running RNArray.}
\usage{
RNArray_UI()
}
\description{
RNArray shiny-based graphical user interface for file selection and running RNArray.
}
\author{
Christina Bligaard Pedersen
}
